+++ pkarr/CHANGELOG.md
@@ -1,8 +1,21 @@
 # Changelog
 
-All notable changes to pkarr client and server will be documented in this file.
+All notable changes to pkarr will be documented in this file.
 
 ## [Unreleased]
+
+### Added
+
+- Add feature `endpoints` to resolve `HTTPS` and `SVCB` endpoints over Pkarr
+- Add feature `reqwest-resolve` to create a custom `reqwest::dns::Resolve` implementation from `Client` and `relay::client::Client`
+- Add feature `tls` to create `rustls::ClientConfig` from `Client` and `relay::client::Client` and create `rustls::ServerCongif` from `KeyPair`.
+- Add feature `reqwest-builder` to create a `reqwest::ClientBuilder` from `Client` and `relay::client::Client` using custom dns resolver and preconfigured rustls client config.
+
+### Changed
+
+- Replace `Settings` with `Config` with public fields.
+
+##  [3.0.0](https://github.com/pubky/mainline/compare/v2.2.0...v3.0.0) - 2024-12-05
 
 ### Added